Improvements to stat scripts
* More HP to player than enemies
* Disabled XP and Level related stuff
* Edited the scene for combat testing
* Added more comments to code
* cleanup

+++ Assets/Scripts/Interact.cs
@@ -14,7 +14,7 @@ public class Interact : MonoBehaviour
     // Roll for attack
     public int AttackRoll()
     {
-        difficulty = (int)gameObject.GetComponent<StatSheet>().DEF;
+        difficulty = gameObject.GetComponent<StatSheet>().DEF;
         roll = new SkillCheck(d, dCount, mod);
         return roll.Roll();
     }
